+++ docs_rtd/changelog.rst
@@ -6,6 +6,7 @@ Changelog
 
 Develop
 -----------------
+* [FEATURE] Add sqlalchemy engine support for `column.most_common_value` metric
 
 0.13.22
 -----------------